requirements.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
django-cors-headers==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
django==5.1.1 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.12" and python_version < "4.0"
drf-yasg==1.21.7 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
python-decouple==3.8 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
setuptools==75.1.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
whitenoise==6.7.0 ; python_version >= "3.12" and python_version < "4.0"
